Fix watch service for projects with many files
This commit fixes two issues in 1.1.0 reported by @francisdb:
  #5)  No files were watched and the log was filled with
       FSEventStreamStart: register_with_server: ERROR: f2d_register_rpc() => (null) (-21)

  #11) During long recompilation stages, file updates were lost by the
       service but would be picked up after compilation completed

The first issue was because we previously were creating a WatchKey for
every directory that was registered with the WatchService. This was a
bad idea because the apple file system events are recursive, so if you
have a watcher on the directory `foo`, it is redundant to have a watcher
on `foo/bar`. Initially to fix this, I checked when we register a new
path if it already covered by a parent path that is being watched. If it
was, then I didn't register it. If a new path was the parent of a
previously registered path, then we stopped listening on the child path.
This solved #5, but caused #11 (which was found by @francisdb using a
snapshot build).

I created issue #11 by keeping only a single WatchKey that covered all
of the subdirectories in the path. This caused two performance issues:

1) Since we reported the root path back to SBT, it had to check all of
   the files in the root path to see which changed, this caused SBT to
   be very slow once it received a file event.

2) When compiling many files, the WatchKey on the root path could fill
   its event buffer (generally with .class files), causing source
   modification events to be dropped.

A naive solution to the second problem was to just increase the buffer
size, which did make the issue go away, but was unsatisfying. Instead of
settling for this, I now store all of the WatchKeys for subdirectories
that have been registered within a path. When a change occurs in the
subdirectory B in root path A, it is handled in the callback for A, but
the WatchKey returned to SBT is for B.