Add tests
It was quite shameful that I hadn't added any tests yet. In this commit,
I add a simple test that an overflow on one watched directory does not
bleed into another. I also added a bunch of basic infrastructure to make
the service more testable. In particular, I added the onOffer callback
function to the MacOSXWatchService. This allows the application to get
notified whenever a new Key