Add flags
The previous implementation didn't really use the event stream flags. As
a result, it was recursively scanning each updated directory for files,
even though the callback (usually) returns the exact path in which a
child file has changed.

I added all of the flags, even though we are only using a few of them.