Refactor MacOSXWatchService
There was a lot of logic that existed in the MacOSXWatchKey that really
belonged in the service. The service is aware of all of the keys and can
find the key to which a file event corresponds. This makes some
subsequent work cleaner.

I also realized that I can now immediately create the run loop thread
since it now waits for a signal before calling CFRunLoopRun().