Improve startup performance
It turns out that it is quite slow to recursively scan large
directories, which can lead to long start up times for continuous builds
for large projects. To fix this, I add some additional concurrency to
the WatchService. In particular, we do all directory traversals on a
background thread and post all of the file system callbacks onto this
same thread.

Because I added additional concurrency, I had to add some additional
callback functions to the service to make it testable. In particular,
the service can now generate callbacks when a key is registered (this is
deferred until after the file system is completed, if necessary), when a
a key is offered to the service (for poll) and finally whenever a file
event is created. I also generalized some of the test cases a bit to
take advantage of these callbacks in a more generic way.

It also is slow, but non-critical, to run the loop that cleans up
WatchKeys that may no longer need FSEventStreamRefs. To prevent that
cleanup from interfering with building the file mapping, I run those on
a separate thread.