Add DirectoryWatcher
The initial implementation of the MacOSXWatchServicePlugin was based on
a port of the https://github.com/takari/directory-watcher project which
used the jna to hook into apple file system events. This directory
watcher largely mimicked the java WatchService api. While this api is
fine, the user must poll the service for events. I prefer to have the
service accept a callback that fires whenever there are events. This is
a more reactive service, which was more suitable to my needs. It also
can be wrapped by a cache to implement a WatchService with minimal
ceremony.

I was also motivated to switch to the jni from the jna. I was seeing
crashes in the jna in some object Finalizers. I also wanted more fine
grained control of the system api calls that I was making. While I could
have done this with jna, it was straightforward with jni. The jni
implementation also performs better and I have yet to trigger a crash
while using it.

I also discovered that when registering a directory to the watch
service, there is a flag that makes the service return events for the
actual files that are modified, rather than an event for the parent
directory. This is quite nice because it means that we don't have to
traverse the file system to trigger events. It can increase the volume
of events in some situations, so the user may prefer to disable the file
events in favor of directory only events. In this case, the events are
delivered recursively for the base directory that is registered. This
means that if an event happens in /foo/bar, the event will be for /foo,
which means that it would be necessary to check all of the files that
are children of /foo to determine to which file the event corresponds.

All of the flags available to the file system create api are now exposed
to the user.