Switch to directory watcher
Now that I have added the stand alone directory watcher, I can
re-implement the MacOSXWatchService using it. This makes the
MacOSXWatchService service implementation simpler because it no longer
has to track whether or not to create a watch stream for each directory
(or remove old streams if a parent directory is added to the watch
list).