Fix service event creation
I realized that many events were being ignored because they didn't have
exactly the expected flags. From now on, ENTRY_CREATE is going to be
best effort, but since the file system isn't cached, it can be fooled by
file move events that should be treated as creates (such as writing a
vim file for the first time). All events that are neither creation or
deletion events will now be reported as modify events.