Add improved equality check for Traversables
In testing, we want to assert that all of the elements of two
traversables are the same. While the ==> method in utest does the right
thing when the two collections are of the same type, it will perform an
equality check when they aren't the same type (which is guaranteed to
fail). To fix this, I add the === extension method to any container
class that requires that each container have a common super type.

Bonus: add dir and file to the names of the files created by
withTempDirectory and withTempFile. This makes debugging easier.