Improve thread model in DirectoryWatcher
There was previously an odd mix of threads and executor services. To
improve coherency, I switched to exclusively using ExecutorServices
(with a single thread) instead of manually managing a thread. I also
reworked on which thread certain tasks ran. Now the callback executor is
only used for handling the callback that is passed in. The register and
unregister methods no longer run on an executor, but this requires that
we use locks where we modify mutable data structures or interact with
the file system api.