Add file system cache framework
SBT does not maintain a cache of source files. This means that every time
that it needs to access the list of files, it must traverse the file
system. When a project only has a few dozen files, this is quick enough
that the user doesn't really notice. However, when a project has hundreds of
files, traversing the file system becomes slow O(1 second) on my mac
with 20000 files. Moreover, SBT sometimes makes multiple passes of the
source list, which can really push out the delay before anything happens
on a new command.

To improve the performance of listing the source files, I add a caching
framework. First I add a Directory class which maintains the cache for a
particular directory. The directory is recursive, so we need only one
Directory to handle all of the subdirectories. Each directory contains a
collection of files and subdirectories. To retrieve all of the files and
directories that are children of the parent, we recursively traverse all
of the subdirectories when calling list (when the recursive flag is
set).

The FileCache wraps a DirectoryWatcher* with a collection of
directories. Whenever we call the list method, we check if the cache is
already watching the directory and, if not, we add it to the watched
directories, otherwise we just list the files from the cached directory.

I also add a NoCache class which doesn't cache the files, but does add
the listed directory to the DirectoryWatcher.

It is obviously a bit risky to rely on file update events to maintain
the state of the cache, but the cache can be invalidated by calling
reload.

* It actually wraps up to two directory watchers. The reason for this is
that I was concerned about cache consistency, so I thought it might be a
good idea to have a low latency file watcher along with a high latency
directory watcher (for anti-entropy). This hasn't really been needed as
the low latency file watcher seems to work well, but I thought it was a
good idea nonetheless.