Add new continuous execution implementation
Right now, there are numerous issues with continuous builds in sbt. For
example, it doesn't correctly cache the WatchService between builds,
which I believe leads to missed file events because the WatchState is
not linked to the correct WatchService. Rather than wait for the
upstream changes, I decided to go ahead and reimplement the '~' command.

The new implementation is more reactive than the old. Rather than
polling periodically for file events and user input, I create a new
thread for user input (that just does a blocking read) and block on a
shared queue that receives input and file events. File events trigger a
new build, while input events stop the continuous task. To revert to the
old watch service implementation, the user can set the
useDefaultWatchService flag to true.

The FileEvents are managed by a user specified FileCache. Hopefully, the
default options are sufficient, but the latency and whether or not to
use a file watcher, a directory watcher, or both is user configurable.

I also made it possible to override the source file list task. This task
is quite slow in large projects O(1 second) and can make the continuous
build feel sluggish. It can (by default) use the same FileCache as the
continuous execution command, but if the useDefaultSourceList flag is
set then it will fall back on the default task.

I had to move the MacOSXWatchServicePlugin into the
com.swoval.watchservice package and moved the file as well.