Fix base directory handling
The new file cache was not previously watching for files in the project
base directory. This meant that if you had the plugin added globally,
you could not actually load your project if it had any dependencies
(e.g. project/Dependencies.scala). To fix this, I add the base directory
to the list of directories that the file cache watches. Adding the base
directory caused continuous execution thrashing though due to sbt 3439.
Unfortunately, the fix that added the recursive boolean to Source only
impacted the default watch service. In my case, I was still getting
tripped up by the recursive search in the base directory because the
recursive flag only applied to the getUnfilteredPaths method (which we
decidedly do not want to call). To fix this, I had to override the
watchSources to replace the base directory Source with a Source that
only includes files in the base directory (and not recursively). It
would be nice to fix sbt io upstream, but I'm not holding my breath.

As a bonus, I fixed the assignment of the includeFilters. For some
reason doing this with state sort of worked, but it's cleaner to assign
them in the project settings.