Split apple-file-system into its own project
I want to update my plugin to work with linux, but I can't build the
apple file system stuff on a linux box, so I'm splitting it out into its
own project. Because the file system api is scala version independent, I
decided to port it to java so that it would work with any scala version
(provided that the jdk is >= 8).