Project overhaul
This is an irresponsibly large commit that completely re-works this
project. It splits the project into three main projects:

1) apple-file-system -- this provides a small api to hook in to apple
   file system events

2) files -- utilities for traversing and caching the contents of the
   file system

3) plugin -- this uses projects 1 and 2 to replace the default watch
   service

Projects 1 and 2 are cross projects that run on the jvm and scala.js (it
would be straightforward to extend to scala native once thread support
is enabled). The plugin only runs on the jvm for obvious reasons.

There isn't really a lot of new functionality in this commit, though I
did fix a few bugs in Directory.scala and ExtensionFilter.scala
(https://github.com/swoval/MacOSXWatchService/issues/33). Getting
everything to work in javascript was a bit of a challenge. I had to
write a node native extension to implement the FileSystemApi. I also had
to rework all of the tests to use Promise/Future since the file system
apis are asynchronous. Since java.io and java.nio are generally not
implemented for scala.js, I also factored out the portions of
java.io.File/java.nio.file.Path that I was using into
com.swoval.files.Path and implement it for scala.js.

Notes:
  - The apple-file-system jvm library is written in pure java, allowing
    it to be used in any jvm >= 8 project. Because scala.js only accepts
    scala source files, I convert the java files into scala files using
    scalagen (https://github.com/timowest/scalagen).

  - I also add linux support with the NioDirectoryWatcher. In my
    experience, the default continuous execution implementation feels
    sluggish on all platforms (although it is at least useable on
    linux). I didn't add linux support in scala.js yet. The
    NioDirectoryWatcher for scala.js just has stubs.

  - I moved the logic that cleans up redundant directory streams into
    native c++ code that is shared by both the jvm and node.js
    extensions.