De-duplicate sources
The (un?)managedSource* settings and tasks return Seq, not Set so there
is nothing preventing the user (or sbt internally) from adding the same
file or path multiple times. This causes the plugin to return the same
source file multiple times, which in turn causes compilation to fail. To
fix this, I just de-duplicate by calling distinct on the lists. This
seemed to have little impact on performance.

Should fix #36 (https://github.com/swoval/MacOSXWatchService/issues/36)