Exclude managed sources from watch sources
It is possible (though inadvisable) to generate sources in the unmanaged
source directory. When this happens, continuous builds may fail because
the generated source files are in the watch source list and change
during each build. To prevent this, filter out any managed sources that
are in the unmanaged source directory from the watch sources.