Rework js portion of project
I reworked the js project so that there are no explicit npm
dependencies. To do this, I needed to add a hack where I could access
the global require function, which is not available in any simple way in
scala.js. I create an unlikely dynamic field name ___global and then
strip out the entire $prefix.global on the generated js file. I think in
scala.js 1.0, I would be able to directly access the global scope, but I
couldn't get my build to work with 1.0.0-M2.

I also added npm directories to apple-file-events and files which are
projects from which I can publish npm packages. The files npm project
exports DirectoryWatcher and FileCache classes with apis that only use
native javascript objects. I also updated the install script for
apple-file-events.js to only build the file if the md5sum of the source
files has changed.

With this change, the pre-compiled node binary will be distributed in
the npm package instead of building on the fly when the package is
installed.