Simplify file cache
It was complicated (and unused behavior) to have separate watchers for
directories and files. I had added this for anti-entropy in the case
where file updates where not correctly processed. The idea was that the
higher latency directory watcher could cleanup after the low latency
file watcher, but I don't think this is necessary in practice and it
made the code a lot more complicated.