Don't reuse FileCache
Previously the FileCache had been a global. This worked most of the
time but, when cross building, the current project is closed which
caused the global cache to be closed. Later attempts to list the files
from the cache would then throw an exception. The simplest fix is to
just change the FileCache default from a lazy val to a def. This does
mean, however, that every time we make a new cache, we have to
re-traverse the file system to clear it. To avoid this, we now pass in
the directory set as a parameter to the FileCacheImpl.