Register WatchSources in Continuously
I noticed that watch events weren't being detected in subprojects (e.g.
when I ran ~foo/compile instead of ~compile). This was because
I had previously relied on all of the WatchSources being implicitly
watched by the FileCache due to prior directory listing. This would work
because when we ran a task like ~foo/compile, when the foo/compile task
was run, it would implicitly register all of the source directories with
the cache. The problem is that (for reasons I don't fully understand),
different caches were being used by executeContinuously and the
foo/compile command.

It was a bug that I wasn't doing this before since WatchSources are not
strictly the same as the source directories.

It's sad that I had to use reflection to get the base directory out of
source (at least it wasn't private[this]).

Fixes #45