Refactor project build
I refactored the build to properly package everything in the com.swoval
package. Intellij complains if you put a packaged file in the top level,
so I had to modify the project build.sbt to include all of the scala
files in the com/swoval package because, by default, sbt only includes
scala files in the base directory OR in src/main/scala (which only
confuses intellij futher).

Bonus:
  - bump sbt version (< 1.1.0 has some zinc issues)
  - cleanup dependencies

squash! Refactor project build