Refactor npm install scripts
I decided to stop checking in the swoval_apple_file_system.node files.
Instead, I have sbt generate install scripts for files.js from the
scripts in appleFileEvents.js. The main inspiration was that I realized
that end users may be using different versions of node, so it's
important that the native extension be built against their version. Also
I noticed that the swoval_apple_file_system.node contains a string that
is the absolute path of the source file from which it was compiled. I
assume that string is used for debugging information and I'd prefer that
it mapped to a file on the local file system of the end users rather
than the path where I built the file.

I also removed the caching logic from the install.js script and moved
that logic into Build.scala. SBT will always run node install.js the
first time appleFileEventsJS.compile is run, but subsequent calls will
only call node install.js if either the source or binary files have
changed.