Improve native file system libraries
I noticed that the filesJS tests would occasionally crash and I tracked
it down to some behavior that wasn't thread safe in the node native lib.
Being new to node, I didn't realize that most of the libuv functions are
not thread safe and that the idiomatic way to run work on the main
thread is with uv_async_init / uv_async_send. Using this idiom allowed
me to simplify the node adapter. I also updated the header to not ever
manage the loop thread and I made the event and stop stream callbacks
take the handle lock as arguments to ensure that they have the lock
while they're called.

Finally, I renamed the native sources files with a more normalized
scheme.