Make build cross platform
This updates the build to be cross platform for scala 2.10, 2.11 and
2.12. I also make the build cross sbt compatible so that either sbt 0.13
or 1.x can be used. The one catch is that releases must be made from
1.1.1 because sbt-doge does not quite provide the same cross building
functionality that sbt 1.x does.

To make sources compatible, I had to remove some SAM code, modify some
macro definitions. The 2.10 build doesn't work for the js projects
(there is no 2.10 version of node.fs, unfortunately) or reflect (this
could probably be made to work, but I don't care enough to do so at the
moment).