Minor changes to FileEvents
This should be a functional no-op, but it adds the isCreated() method to
FileEvent (which is only used by isNewFile at the moment). It also stops
trying to load the native library from the system. Right now the native
library is only distributed inside of the apple-file-events jar, so
there is no point in looking in the system for the library (this was a
legacy from when I copied the body from the generated code from the
jni-native plugin).