Add anti-entropy to FileCache
I found that it was impossible to get the FileCacheTest to reliably pass
because sometimes the cache would receive two events for the same file
when the file was renamed. To fix this, I add a small amount of
anti-entropy to the cache. Now, when an event comes in for a file, the
cache schedules processing of the event. If another event comes in
before the scheduled work, the event is ignored and will be implicitly
handled by the previously scheduled work.