Switch to java 7 source compatibility
In order to ensure that the plugin should work for most sbt 0.13 users, I
downgraded apple-file-events to only use java 7 compatible apis. My
previous company was still stuck on java 7 and I'm sure many other are
too. I'm not worried about the improved lambda performance in java 8 at
the moment.  In the future, there could be a java 8 and a java 7 version
of the package, but this seems good for now.