Add recursive flag to FileCache.register
I noticed when testing in linux that a number of the target paths were
being watched because the recursive source was not being handled
correctly in the base directory. Because the recursive flag was always
passed in as true to the watcher, it would recursively traverse the base
directory and add unwanted target directories to the watch. They
wouldn't trigger a build, but added inotify nodes and made the logs more
noisy.