Fix FileCache lifecycle
In bb4afc14a34, I tried to fix a bug where a closed FileCache would be
re-opened. This didn't quite work as well as I'd hoped and lead me down
an sbt setting/task rabbit hole. To make a long story short, the safest
way to manage the life cycle of the cache is to make it a var that is
initialized on project load and closed on project unload. I also
switched the cache to a TaskKey both because I need to make a new one
every time the project is re-loaded and also because there is some
bizarre stuff with setting key evaluation that can cause vals to be
evaluated multiple times for the same object (I don't really understand
what is going on with the classloaders).