Fix segfaults
It isn't safe to close the file event handle on an arbitrary thread. To
work around this, I make the close method shutdown the run loop
asynchronously ON the run loop. While the shutdown is happening the
close method blocks on a condition variable. Once the close method
returns, it should be safe to delete the handle because no further
callbacks will be submitted to the run loop.

There also was a race condition where close could be called before the
run loop thread had finished starting up. This also caused crashes.