Fix race condition
Prior to this change, a number of callbacks could be invoked on the
subdirectories and files before these new paths were added to the
directory. This had previously caused spurious text failures that I was
able to workaround by scheduling the FileCacheTest callback to run on a
different thread, but this properly fixes the race condition.