Close node apple file system handle immediately
I realized that because callbacks are scheduled to run on the main
thread and we cancel the callbacks as part of closing the handle, there
is no need to defer the close to process_callback.

I also found that the js tests would frequently hang. This was resolved
by making the tests less async which required a bit more verbosity
because I had to add some manual cleanup code that wasn't required with
the usingAsync method.

Bonus: don't double close