Upgrade plugins
I found some performance issues with the scala.js bundler plugin and the
scala js plugin itself. I fixed the issues on my local machine, but
it may take a while for the changes to appear upstream. To get around
this, I make these plugin versions dynamically settable with a system
property.