Invoke callback on directory creation
I noticed that the FileCacheTest.files.register.addmany test would
frequently fail on linux and the reason was that when a new directory
was created, no callback would be invoked for the new directory.

I also realized that the order of the callbacks was unpredictable
because executor.schedule doesn't really have any ordering guarantees.
To work around this, I added a queue of jobs to run and schedule work
based on the state of the queue. I added a test to ensure that, for
example, when a new directory is created, the callback for the directory
always runs before the callback for any files in that new directory.