Re-implement file utiltitis in java
This is a sprawling commit that more or less rewrites the entire files
project. The motivation was that I wanted the file-utilities library to
be a java only project. The library barely used any scala features and
by implementing it in java, I reduced the jar size (by 4X) and made it
possible to use the library in any jvm language without including the
scala standard library. I also don't have to cross publish for different
scala versions, which is nice in light of the moving target that is
scala 2.13 and scala 3.

This is more than just a simple rewrite. I changed the api of Directory and
FileCache so that instead of just storing Path instances, they now store
a cache entry that has a user defined data value in addition to the
path. In the case when the data value is just the path itself, the
semantics are more or less identical as before. Moreover, I updated the
FileCache callback mechanics to now take an Observer instance where the
observer provides three different callbacks that run in the case of file
create, delete and update events. The creation callback will receive the
new cache entry for the created file, the deletion callback will receive
the old cache entry for the now deleted file and the update callback
will receive both the old and new values.

I also removed the swoval Path class and am just using
java.nio.file.Path everywhere. This makes interoperability easier. It
did mean that I had to implement a good fraction of java.nio.file and
java.io.File for JS. This may be useful as a standalone module.

I consolidated apple-file-events and files into a single project. The
apple-file-events jar was ~20Kb before and the file-utilities jar was
~180Kb. The new combined file-utilities jar is ~65Kb.

In the java rewrite, I had two big restrictions: I wanted to keep it
java 7 compatible (so no lambdas could be used) and I wanted code gen to
work. There are a number of java features that are either incorrectly or
poorly translated by the scalagen tool. The two main issues are java
for loops with ':' and early returns. In the former case, scalagen does
something like translating
for (final Path path : paths) System.out.println(path)
to
import scala.collection.JavaConversions._
for (path <- paths) println(path)
The problem with this is that now JavaConversions are in the global
imports for the file and it makes it difficult to reason about a number
of operations. To fix this I restrict myself to iterator operations so
that I'd write the above to
final Iterator<Path> it = paths.iterator()
while (it.hasNext) System.out.println(it.next())

Scalagen doesn't add early returns so I just had to rewrite very method
to only have a single return. This came up because I was translating a
number of tail recursive methods to iterative methods where an early
return in the base case was convenient. It is not that difficult to work
around.

For the most part, I left all of the existing tests in place. I added
some new tests for Path (which is mostly to check that the js
implementation satisfies the subset of the api that is important to this
project). There were some changes that I had to make because the apis
changed and I also split the AppleDirectoryWatcherTest into two tests
where the main DirectoryWatcherTest also runs on linux.

I added a ton of javadoc so that the majority of public apis are now
documented. I also reworked the code so that as few classes and methods
are as public as needed for a reasonable api.

Finally, to make the project more IDE friendly, I changed the way that the
shared sources work. Intellij does not see the platform specific code
when you are working with a shared sources, which leads to obnoxious red
squiggles and (even worse) non-dismissible auto-complete pop ups. To fix
this, I override the compile task to create symlinks to all of the
shared source files and then exclude the files in the shared directory
from the sources task. This makes developing in the IDE far more
pleasant.