Fix deletion callbacks
I discovered that on linux sometimes I would not always get the watch
event for removed file in a subdirectory if the subdirectory itself was
also deleted around the same time. I had intended to handle this edge
case by returning all of the deleted files in the deleted subdirectory
initially, but somehow it slipped my mind. On OSX, the file events
pretty much always occur before the directory events, so I don't think
this change is as relevant there.