Fix overflow behavior for NioDirectoryWatcher/FileCache
It turns out that it's critically important to handle OVERFLOW events
that come from the NioDirectoryWatcher. This is because we rely on the
file events to ensure that we are watching all of the subdirectories for
a watched path. When an overflow occurs, we may miss directory
creation/removals. The remedy is that when we detect an overflow, we
poll the file system for all of the directories. Using these results, we
register any new directories and unregister any directories that no
longer exist. Moreover, because an overflow is likely in some kind of
mass event, like `rm -r`, we repeatedly poll the directory until there
are no changes.

The FileCache has a nearly identical issue, so when the DirectoryWatcher
detects an OVERFLOW, it reports that back to the FileCache, which then
does a similar poll operation to ensure that the directory state matches
the file system. It only stops polling once the cache is stable relative
to the file system.

I wanted to ensure that I tested this logic on OSX, so I added
additional apply methods to FileCache that take in a new
DirectoryWatcher.Factory instance, which can be used to instantiate the
DirectoryWatcher used by the cache. This allows us to continue using the
AppleDirectoryWatcher on OSX, but also test the NioDirectoryWatcher
logic. The DefaultFileCacheTest already uses nio on linux, so the
NioFileCacheTest only runs on OSX.