Add overflow test for NioDirectoryWatcher
This adds a test that should ensure that an overflow occurs in
NioDirectoryWatcher. The NioFileCacheTest should often trigger an
overflow, but it isn't guaranteed. I noticed looking at the jacoco
reports that the overflow case wasn't always captured.

This test was a bit tricky to write because the queue size of the
default watch service is quite big (512) and not configurable. To try
and add some determinism, I added the BoundedWatchService class. It
wraps regular WatchService but returns keys that will drop the tail of
the list of events if the list size exceeds the queue size. This allows
me to reliably induce an overflow without creating a massive amount of
files.

The test itself registers a temporary directory with the service. It
then creates 10 subdirectories (which should trigger an overflow because
the queue size is only 2 and creating directories is relatively fast).
We then verify that the DirectoryWatcher detects file events in each of
the created subdirectories.

The test isn't really relevant in scala.js because the scala.js
NioDirectoryWatcher does not use a WatchService, but to make things link
and compile, I had to implement more of the java.nio.file package. Long
term, I may want to actually implement WatchService for scala.js.