Add Registerable interface
The interface of java.nio.file.WatchService does not include a method
for registering a watchable with the service. This is unfortunate
because Path.register(WatchService) only works on the jdk provided watch
service. As a result, anyone using the MacOSXWatchService vs. the
default LinuxWatchService would have to use runtime instanceof checks to
register a path with the service.