Add QuickList
This commit primarily adds the QuickList class that provides list
methods that can be significantly faster than using Files.walkFileTree
or rolling your own with Files.list or File.listFiles. I added native
implementations for osx (64 bit only) and linux and windows (32 bit and
64 bit). The native code is very simple and mainly just delegates to the
underlying system calls to list a directory. It also translate system
error codes to swoval specific values with the same semantics so that
proper exceptions may be thrown when errors are encountered.

The reason that QuickList is worth having is that the stat system call
is very expensive. I made a simple benchmark where I filled a directory
with 5000 files and listed it. Using quicklist, it took roughly 25ms on
my mac, while using Files.walk took roughly 110ms. I then ran separate
benchmark where I just called Files.readAttributes (which eventually
calls the native stat function on the file) and that took roughly 75 ms,
which accounted for most of the difference in performance. My hunch is
that by far the most common file attribute that people actually use is
isDirectory or isFile, both of which can be provided without calling
stat. This is certainly true for the FileCache, which can now fill the
cache more quickly than before.

There was a lot of plumbing I had to add to make this work. I needed to
augment the native makefile to handle the linux and windows builds. I
also had to add the respective
com_swoval_files_NativeQuickLister.cc files (which use completely
different apis to effect the same semantics). I had to overhaul the
NativeLoader so that it would work for different architectures and
platforms. Windows also doesn't let you delete a .dll that is in use, so
I had to add in a hack to unload the .dll before we try to delete it.

I have some residual concern that I may not have gotten all of the
semantics correct of the NativeQuickLister, so I provide a nio based
version. In spite of QuickList being a singleton, the QuickLister
implementation that it uses may be overridden with the
-Dswoval.quicklister system property, which takes a fully qualified
class name of a QuickLister implementation with a no argument
constructor.

I had to add a number of missing java.nio.file implementations to the
scala.js project.

Bonus: add some missing scaladoc