Add windows support
A number of minor things didn't work correctly on windows, especially
for scala.js. To ensure that they work moving forward, I added a
.appveyor.yml and have added an appveyor integration to swoval. Most
notably, symbolic links are a pain to get to work on windows, so I
crudely copy the files on windows instead symlinking them. This is
really ugly, but I doubt there will be much development of this library
on windows (or potentially by anyone but myself).