Fix NioDirectoryWatcher bug
It is important that the NioDirectoryWatcher does all of the processing
for a given event before it invokes the callback. This is because the
semantics are that the receiver may assume that if the recursive flag
was set on registration, then when it receives a directory creation
update, the watcher will be registered with the directory. If this
assumption doesn't hold, then it's possible that the downstream observer
will miss the creation of some files*.

* Even if those file creations aren't observed by the DirectoryWatcher,
the observer can list the newly created directory and treat all existing
files as newly created.