Add support for monitoring symlink targets
When a file cache is watching a directory and that directory contains
symlinks, the watcher will not necessarily detect when the target file
changes. To work around this, I add a new class SymlinkWatcher that can
monitor the target of the symbolic links in the directory. To do this,
I add Options to the cache, with a boolean parameter followLinks (that
is always true at the moment). When followLinks is true and the cache
finds a symlink, it registers the symlink with the new SymlinkWatcher
class. That class keeps track of all of the targets of symlinks that are
being monitored. For each target, it registers the path.

I had to add a good deal of machinery to handle the possibility of a
file loop. When a loop is detected, the cache will report the error to
the user, but the cache should continue to function.

I also introduce a new interface FileType that provides (fast) methods
for checking if a file is a regular file, directory or symlink. The
QuickFile class can now implement this interface. The interface is
useful for any situation where the user may want to do different things
for the different kind of files returned by QuickList.list (including
filtering out unwanted files).

Adding this functionality revealed a number of issues with how I was
registering paths in the NioDirectoryWatcher. This lead to the a number
of changes in the NioDirectoryWatcher which helped make this commit even
bigger.