Properly remove streams
I started seeing occasional fatal errors when running tests because
native code was running on a thread that wasn't registered with the jvm.
The issue was that stopStream was calling back into java. To fix this, I
just moved the stop stream logic into the cleanupFunc and in the
publicly called variant of stopStream, I signal the run loop and block
until the run loop thread cleans up the stream.