Add filters for QuickLister list methods
It can be useful to apply a filter while listing a directory. In
particular, if the user is listing a directory containing many
subdirectories, they may not want to list all of the subdirectories
(which could be slow). Adding a filter, will allow the user to skip the
directories they don't care about. In the case of file filters, there
isn't much point in passing in a filter compared to just filtering out
the unwanted files in the result list. If the filtering is done in
QuickLister.list, the only difference really is that the result list
will be larger, but (more or less) the same number of java objects
should be created either way (since a QuickFile is always made before
filtering and the same file is added to the result list if it's
accepted).